+++ bindings/nostr-js/src/event/mod.rs
@@ -265,4 +265,9 @@ impl JsEvent {
     pub fn as_json(&self) -> Result<String> {
         self.inner.try_as_json().map_err(into_err)
     }
+
+    #[wasm_bindgen(js_name = asPrettyJson)]
+    pub fn as_pretty_json(&self) -> Result<String> {
+        self.inner.try_as_pretty_json().map_err(into_err)
+    }
 }