+++ bindings/nostr-js/src/event/unsigned.rs
@@ -85,6 +85,11 @@ impl JsUnsignedEvent {
         self.inner.try_as_json().map_err(into_err)
     }
 
+    #[wasm_bindgen(js_name = asPrettyJson)]
+    pub fn as_pretty_json(&self) -> Result<String> {
+        self.inner.try_as_pretty_json().map_err(into_err)
+    }
+
     /// Sign an unsigned event
     ///
     /// Internally: calculate event ID (if not set), sign it, compose and verify event.