+++ bindings/nostr-js/src/types/filter.rs
@@ -194,6 +194,11 @@ impl JsFilter {
         self.inner.try_as_json().map_err(into_err)
     }
 
+    #[wasm_bindgen(js_name = asPrettyJson)]
+    pub fn as_pretty_json(&self) -> Result<String> {
+        self.inner.try_as_pretty_json().map_err(into_err)
+    }
+
     /// Set event ID
     pub fn id(self, id: &JsEventId) -> Self {
         self.inner.id(**id).into()