+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 ### Added
 
 * nostr: impl `TryFrom<Vec<Tag>>` for `LiveEvent` ([w3irdrobot])
+* bindings(nostr): expose `as_pretty_json` for some structs ([Yuki Kishimoto])
 
 ### Fixed